+++ RELEASE-NOTES.md
@@ -30,6 +30,7 @@ Initial version with:
 
 ## [0.7.0](https://github.com/kisscool62/io.nagurea.smsupsdk/compare/v0.6.0...v0.7.0) (2022-07-04)
 * Campaigns > Get campaign
+* Campaigns > Get campaign stops
 
 ### Upgrade Steps
 * [NO ACTION REQUIRED]
@@ -39,6 +40,7 @@ Initial version with:
 
 ### New Features
 * Campaigns > Get campaign
+* Campaigns > Get campaign stops
 
 ### Bug Fixes